fix numpy data type of bonds amd ssbonds arryas in molecule

+++ htmd/molecule/molecule.py
@@ -744,8 +744,8 @@ class Molecule:
                     self.__dict__[k] = numpy.zeros(natoms, dtype=self.__dict__[k].dtype)
 
         self.coords = np.atleast_3d(np.array(self.coords, dtype=np.float32))
-        self.bonds = np.array(mol.bonds)
-        self.ssbonds = np.array(mol.ssbonds)
+        self.bonds = np.array(mol.bonds, dtype=np.uint32)
+        self.ssbonds = np.array(mol.ssbonds, dtype=np.uint32)
         self.box = np.array(mol.box)
 
         if self.masses is None or len(self.masses) == 0:
